scripts/comps-sync/requirements.txt
argparse
libcomps
pyyaml
